+++ src/clib/framework/winfw.h
@@ -33,7 +33,7 @@ void ejoy2d_win_init(struct STARTUP_INFO* startup);
 void ejoy2d_win_frame();
 void ejoy2d_win_update(float delta);
 int ejoy2d_win_touch(int x, int y,int touch);
-void ejoy2d_win_gesture(int type, int x1, int y1, int x2, int y2, int state);
+void ejoy2d_win_gesture(int type, float x1, float y1, float x2, float y2, int state);
 void ejoy2d_win_view_layout(int stat, float x, float y, float w, float h);
 
 #endif
